add: useDesktopMode hook, add cleanup function in useEffect
